+++ src/page/index.ts
@@ -1,0 +1,3 @@
+export { getPageInfo } from './getPageInfo';
+export { makePageElement } from './makePageElement';
+export { getPageComponents } from './getPageComponents';
